Cleanup onroad UI
