+++ selfdrive/ui/mici/onroad/complication.py
@@ -1,25 +1,28 @@
-import pyray as rl
-import numpy as np
 import time
+
+import numpy as np
+import pyray as rl
+
 from openpilot.common.constants import CV
+from openpilot.common.params import Params
+from openpilot.selfdrive.ui.bp.lib.ui_debug_logger import bp_ui_log
 from openpilot.selfdrive.ui.ui_state import ui_state
-from openpilot.system.ui.lib.application import gui_app, FontWeight
+from openpilot.system.ui.lib.application import FontWeight, gui_app
 from openpilot.system.ui.lib.multilang import tr
 from openpilot.system.ui.lib.text_measure import measure_text_cached
 from openpilot.system.ui.widgets import Widget
-from openpilot.common.params import Params
-from opendbc.car import structs
 from openpilot.sunnypilot import IntEnumBase
-from openpilot.selfdrive.ui.bp.lib.ui_debug_logger import bp_ui_log
+from opendbc.car import structs
 
 FONT_SIZE = 68
 DIST_FONT_SIZE = 55
 TIME_FONT_SIZE = 50
 UNIT_FONT_SIZE = 24
 WIDTH = 80
-COLOR_DELTA_MS = 4.5  # ~ 10MPH
+COLOR_DELTA_MS = 4.5
 SHADOW_DEPTH = 3
-DELAY = 3.0 #seconds to remove last lead car speed
+DELAY = 3.0
+
 
 class ComplicationType(IntEnumBase):
   off = 0
@@ -28,6 +31,7 @@ class ComplicationType(IntEnumBase):
   lead_car_dist = 3
   lead_car_time = 4
 
+
 class MiciComplication(Widget):
   def __init__(self):
     super().__init__()
@@ -35,21 +39,19 @@ class MiciComplication(Widget):
     self.vRel: float = 0.0
     self._font_bold: rl.Font = gui_app.font(FontWeight.BOLD)
     self._color = np.array([255, 255, 255], dtype=float)
-    self._slower_color = np.array([255,  160,  0], dtype=float) #orangish
-    self._faster_color = np.array([0, 255,  0], dtype=float) #green
+    self._slower_color = np.array([255, 160, 0], dtype=float)
+    self._faster_color = np.array([0, 255, 0], dtype=float)
     self._font_color: rl.Color = rl.Color(255, 255, 255, 180)
     self._car_state = None
     self._render_type = 1
     self._last_active_time = 0.0
-
     self.params = Params()
 
   def _update_state(self):
-     self._render_type = self.params.get("mici_complication")
-     bp_ui_log.state("MiciComplication", "render_type", self._render_type)
+    self._render_type = self.params.get("mici_complication")
+    bp_ui_log.state("MiciComplication", "render_type", self._render_type)
 
   def _render(self, rect: rl.Rectangle) -> None:
-    """Draw the first lead vehicle speed and unit."""
     if self._render_type == ComplicationType.off:
       return
 
@@ -63,21 +65,17 @@ class MiciComplication(Widget):
     has_lead_one = self._lead_one.status if self._lead_one else False
     self._render_lead_indicator = self._radar_state is not None and has_lead_one and in_gear
 
-    match self._render_type:
-      case ComplicationType.lead_car_speed:
-        self._render_lead_speed(rect)
-      case ComplicationType.speed:
-        # BluePilot: Respect "Speedometer: Hide from Onroad Screen" (HideVEgoUI) from Visuals.
-        # Read param directly for immediate response (ui_state.hide_v_ego_ui refreshes every 5s).
-        if not self.params.get_bool("HideVEgoUI"):
-          self._render_current_speed(rect)
-      case ComplicationType.lead_car_dist:
-        self._render_lead_dist(rect)
-      case ComplicationType.lead_car_time:
-        self._render_lead_time(rect)
+    if self._render_type == ComplicationType.lead_car_speed:
+      self._render_lead_speed(rect)
+    elif self._render_type == ComplicationType.speed:
+      if not self.params.get_bool("HideVEgoUI"):
+        self._render_current_speed(rect)
+    elif self._render_type == ComplicationType.lead_car_dist:
+      self._render_lead_dist(rect)
+    elif self._render_type == ComplicationType.lead_car_time:
+      self._render_lead_time(rect)
 
-
-  def _render_lead_speed(self,rect: rl.Rectangle):
+  def _render_lead_speed(self, rect: rl.Rectangle):
     if self._render_lead_indicator:
       self._last_active_time = time.monotonic()
       speed_conversion = CV.MS_TO_KPH if ui_state.is_metric else CV.MS_TO_MPH
@@ -89,16 +87,15 @@ class MiciComplication(Widget):
       delay_time = time.monotonic() - self._last_active_time
       if delay_time > DELAY:
         return
-      else:
-        fade_ratio = 1.0 - (delay_time / DELAY)
+      fade_ratio = 1.0 - (delay_time / DELAY)
 
     v_delta = np.clip(self.vRel, -COLOR_DELTA_MS, COLOR_DELTA_MS)
     if v_delta <= 0:
-        t = (v_delta + COLOR_DELTA_MS) / COLOR_DELTA_MS
-        result = (1 - t) * self._slower_color + t * self._color
+      t = (v_delta + COLOR_DELTA_MS) / COLOR_DELTA_MS
+      result = (1 - t) * self._slower_color + t * self._color
     else:
-        t = v_delta / COLOR_DELTA_MS
-        result = (1 - t) * self._color + t * self._faster_color
+      t = v_delta / COLOR_DELTA_MS
+      result = (1 - t) * self._color + t * self._faster_color
 
     color = result.astype(int)
     self._font_color = rl.Color(color[0], color[1], color[2], int(220 * fade_ratio))
@@ -128,9 +125,9 @@ class MiciComplication(Widget):
     rl.draw_triangle_fan(chevron, len(chevron), rl.Color(201, 34, 49, int(150 * fade_ratio)))
 
   def _render_current_speed(self, rect: rl.Rectangle) -> None:
-    # BluePilot: Respect "Speedometer: Hide from Onroad Screen" (HideVEgoUI) from Visuals
     if ui_state.hide_v_ego_ui:
       return
+
     self._font_color = rl.Color(255, 255, 255, 220)
     shadow_color = rl.Color(0, 0, 0, 180)
 
@@ -152,7 +149,7 @@ class MiciComplication(Widget):
     unit_pos.y -= SHADOW_DEPTH
     rl.draw_text_ex(self._font_bold, unit_text, unit_pos, UNIT_FONT_SIZE, 0, self._font_color)
 
-  def _render_lead_dist(self,rect: rl.Rectangle):
+  def _render_lead_dist(self, rect: rl.Rectangle):
     if self._render_lead_indicator:
       self._last_active_time = time.monotonic()
       self.dist = self._lead_one.dRel
@@ -163,8 +160,7 @@ class MiciComplication(Widget):
       delay_time = time.monotonic() - self._last_active_time
       if delay_time > DELAY:
         return
-      else:
-        fade_ratio = 1.0 - (delay_time / DELAY)
+      fade_ratio = 1.0 - (delay_time / DELAY)
 
     self._font_color = rl.Color(255, 255, 255, int(220 * fade_ratio))
     shadow_color = rl.Color(0, 0, 0, int(180 * fade_ratio))
@@ -186,7 +182,7 @@ class MiciComplication(Widget):
     unit_pos.y -= SHADOW_DEPTH
     rl.draw_text_ex(self._font_bold, unit_text, unit_pos, UNIT_FONT_SIZE, 0, self._font_color)
 
-  def _render_lead_time(self,rect: rl.Rectangle):
+  def _render_lead_time(self, rect: rl.Rectangle):
     if self._render_lead_indicator and self._lead_one.vRel > 0:
       self._last_active_time = time.monotonic()
       self.dist = self._lead_one.dRel
@@ -196,8 +192,7 @@ class MiciComplication(Widget):
       delay_time = time.monotonic() - self._last_active_time
       if delay_time > DELAY:
         return
-      else:
-        fade_ratio = 1.0 - (delay_time / DELAY)
+      fade_ratio = 1.0 - (delay_time / DELAY)
 
     self._font_color = rl.Color(255, 255, 255, int(220 * fade_ratio))
     shadow_color = rl.Color(0, 0, 0, int(180 * fade_ratio))
