+++ selfdrive/ui/mici/onroad/model_renderer.py
@@ -8,6 +8,7 @@ from openpilot.common.filter_simple import FirstOrderFilter
 from openpilot.selfdrive.locationd.calibrationd import HEIGHT_INIT
 from openpilot.selfdrive.ui.ui_state import ui_state, UIStatus
 from openpilot.selfdrive.ui.mici.onroad import blend_colors
+from openpilot.bluepilot.ui.lib.bp_shaders import draw_rainbow_polygon
 from openpilot.system.ui.lib.application import gui_app
 from openpilot.system.ui.lib.shader_polygon import draw_polygon, Gradient
 from openpilot.system.ui.widgets import Widget
@@ -344,7 +345,10 @@ class ModelRenderer(Widget, ModelRendererSP):
     self._blend_filter.update(int(allow_throttle))
 
     if ui_state.rainbow_path:
-      self.rainbow_path.draw_rainbow_path(self._rect, self._path)
+      # BluePilot: use the migrated rainbow shader path for MICI
+      v_ego = sm['carState'].vEgo
+      rainbow_v = np.clip(v_ego, 2.5, 35) / 30
+      draw_rainbow_polygon(self._rect, self._path.projected_points, rainbow_v=rainbow_v)
       return
 
     path_pts = self._path.projected_points + np.array([self._rect.x, self._rect.y], dtype=np.float32)
